requirements.txt
django-environ
django-filter
djangorestframework
black